First pass at log-chasing version of run_sas_program.py.
I'm part of a consortium of orgs that frequently distributes SAS programs to one another to run.  In order to make things nice and clear, these programs drop datasets, reports etc. in "to-stay" and "to-go" folders.  That is, some things are intended to be sent back to the originating site, and others are not.  Because most of these programs want to see the log, they redirect it to a file in the to-go folder.  That screws up the simpler version of the log-checker that's in this package.  Thus, this elaboration.